feat: Implement "press back again to exit" on HomeScreen
This commit introduces a "press back again to exit" functionality on the `HomeScreen`.

When the back button is pressed once, a `SnackBar` is displayed prompting the user to press back again to exit. If the back button is pressed again within 2 seconds, the app will exit.

This is achieved by wrapping the `Scaffold` in a `PopScope` widget and handling the `onPopInvoked` callback.

Additionally, a minor type correction was made in `_todayGroupsList` method signature, changing `List<StudentItemUiState>` to `List<GroupItemUiState>`.