Populate student/parent name from contacts when phone number is selected
When selecting a parent's phone number from contacts in the "Add Student" screen, the student's name field will now also be populated with the contact's name.

This also increments the Android app version:
- `versionCode` to 8
- `versionName` to 1.7

The `AppPhoneInputTextFieldWidget` now accepts an `onContactSelected` callback which receives both the phone number and the contact's name. This callback is utilized in `AddStudentScreen` to set the respective controllers.